Basic service back blades

+++ try/database/migrations/2023_02_27_161305_create_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('title', 50);
-            $table->decimal('min', 2)->unsigned();
+            $table->bigInteger('min')->unsigned();
             $table->decimal('price', 5, 2)->unsigned();
             $table->unsignedBigInteger('salon_id');
             $table->foreign('salon_id')->references('id')->on('salons');
